# packages/contracts/openapi/v1.yaml
...
openapi: 3.1.0
info:
  description: |-
    Version 1 of the Chaya 02 HTTP API. Breaking changes require a new /api/v2 document.

    Generated from the backend's controllers; packages/contracts/openapi/v1.yaml is the committed output and OpenApiContractTest fails when it drifts.

    Roles (x-chaya-roles) under /api/v1/venues/{venueId}/ are the roles the caller holds at that venue. A resource outside the caller's organization or venue answers 404, never 403.

    Errors are RFC 9457 problems with a stable machine-readable `code`.
  title: Chaya 02 API
  version: v1
servers:
- description: The API's own origin
  url: /
paths:
...
  /api/v1/venues/{venueId}/captures:
    get:
      operationId: captureList
      parameters:
      - in: path
        name: venueId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CaptureView"
          description: OK
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Malformed request: an unparseable parameter or body, or a\
            \ failed validation (code BAD_REQUEST or an operation-specific code)"
        "401":
          description: "Missing, invalid or expired credentials (empty body, WWW-Authenticate\
            \ header)"
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The caller's roles do not allow this operation (empty body),\
            \ or an operation-specific refusal (problem with a code)"
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Not found, or not visible to the caller's organization or\
            \ venue (code NOT_FOUND)"
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Rate limit exceeded (code RATE_LIMITED); see the Retry-After
            header
        "503":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "A dependency is unavailable (code DATABASE_UNAVAILABLE, STORAGE_UNAVAILABLE\
            \ or AUTHENTICATION_UNAVAILABLE); the request was not completed"
      security:
      - bearerAuth: []
      summary: List capture sessions
      tags:
      - capture-controller
      x-chaya-roles:
      - ADMIN
      - VENUE_MANAGER
      - OPERATOR
    post:
      operationId: captureCreate
      parameters:
      - in: path
        name: venueId
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCapture"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaptureView"
          description: Created
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Malformed request: an unparseable parameter or body, or a\
            \ failed validation (code BAD_REQUEST or an operation-specific code)"
        "401":
          description: "Missing, invalid or expired credentials (empty body, WWW-Authenticate\
            \ header)"
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The caller's roles do not allow this operation (empty body),\
            \ or an operation-specific refusal (problem with a code)"
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Not found, or not visible to the caller's organization or\
            \ venue (code NOT_FOUND)"
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Rate limit exceeded (code RATE_LIMITED); see the Retry-After
            header
        "503":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "A dependency is unavailable (code DATABASE_UNAVAILABLE, STORAGE_UNAVAILABLE\
            \ or AUTHENTICATION_UNAVAILABLE); the request was not completed"
      security:
      - bearerAuth: []
      summary: Create a capture session (status CREATED)
      tags:
      - capture-controller
      x-chaya-roles:
      - ADMIN
      - VENUE_MANAGER
      - OPERATOR
  /api/v1/venues/{venueId}/captures/{captureId}:
    delete:
      description: "Raw media, frames, PII staging, reconstruction artifacts, versions,\
        \ detections and their embeddings (docs/privacy-erasure.md). Admin or the\
        \ venue's manager. cascade=true also erases re-scans derived from it. Idempotent."
      operationId: erasureEraseCapture
      parameters:
      - in: path
        name: venueId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: captureId
        required: true
        schema:
          type: string
          format: uuid
      - in: query
        name: cascade
        required: false
        schema:
          type: boolean
          default: false
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErasureView"
          description: "Erasure record, COMPLETED"
        "202":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErasureView"
          description: "Erasure record, OBJECTS_PENDING: rows gone, objects still\
            \ being deleted"
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Malformed request: an unparseable parameter or body, or a\
            \ failed validation (code BAD_REQUEST or an operation-specific code)"
        "401":
          description: "Missing, invalid or expired credentials (empty body, WWW-Authenticate\
            \ header)"
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: ERASURE_NOT_PERMITTED
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Not found, or not visible to the caller's organization or\
            \ venue (code NOT_FOUND)"
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: ERASURE_HAS_DEPENDENTS (repeat with cascade=true)
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Rate limit exceeded (code RATE_LIMITED); see the Retry-After
            header
        "503":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "A dependency is unavailable (code DATABASE_UNAVAILABLE, STORAGE_UNAVAILABLE\
            \ or AUTHENTICATION_UNAVAILABLE); the request was not completed"
      security:
      - bearerAuth: []
      summary: Erase a capture and everything derived from it
      tags:
      - erasure-controller
      x-chaya-roles:
      - ADMIN
      - VENUE_MANAGER
      - OPERATOR
      - VIEWER
    get:
      operationId: captureGet
      parameters:
      - in: path
        name: venueId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: captureId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaptureView"
          description: OK
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Malformed request: an unparseable parameter or body, or a\
            \ failed validation (code BAD_REQUEST or an operation-specific code)"
        "401":
          description: "Missing, invalid or expired credentials (empty body, WWW-Authenticate\
            \ header)"
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The caller's roles do not allow this operation (empty body),\
            \ or an operation-specific refusal (problem with a code)"
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Not found, or not visible to the caller's organization or\
            \ venue (code NOT_FOUND)"
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Rate limit exceeded (code RATE_LIMITED); see the Retry-After
            header
        "503":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "A dependency is unavailable (code DATABASE_UNAVAILABLE, STORAGE_UNAVAILABLE\
            \ or AUTHENTICATION_UNAVAILABLE); the request was not completed"
      security:
      - bearerAuth: []
      summary: Inspect a capture session
      tags:
      - capture-controller
      x-chaya-roles:
      - ADMIN
      - VENUE_MANAGER
      - OPERATOR
  /api/v1/venues/{venueId}/captures/{captureId}/complete-upload:
    post:
      operationId: captureCompleteUpload
      parameters:
      - in: path
        name: venueId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: captureId
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteUpload"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaptureView"
          description: OK
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Malformed request: an unparseable parameter or body, or a\
            \ failed validation (code BAD_REQUEST or an operation-specific code)"
        "401":
          description: "Missing, invalid or expired credentials (empty body, WWW-Authenticate\
            \ header)"
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The caller's roles do not allow this operation (empty body),\
            \ or an operation-specific refusal (problem with a code)"
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Not found, or not visible to the caller's organization or\
            \ venue (code NOT_FOUND)"
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Not ready: UPLOADS_NOT_SETTLED, INSUFFICIENT_MEDIA, INVALID_CAPTURE_TRANSITION,\
            \ STORAGE_INTEGRITY"
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Rate limit exceeded (code RATE_LIMITED); see the Retry-After
            header
        "503":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "A dependency is unavailable (code DATABASE_UNAVAILABLE, STORAGE_UNAVAILABLE\
            \ or AUTHENTICATION_UNAVAILABLE); the request was not completed"
      security:
      - bearerAuth: []
      summary: Finish the upload phase (READY_FOR_PROCESSING)
      tags:
      - capture-controller
      x-chaya-roles:
      - ADMIN
      - VENUE_MANAGER
      - OPERATOR
...
  /api/v1/venues/{venueId}/captures/{captureId}/media:
    get:
      operationId: captureListMedia
      parameters:
      - in: path
        name: venueId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: captureId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MediaView"
          description: OK
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Malformed request: an unparseable parameter or body, or a\
            \ failed validation (code BAD_REQUEST or an operation-specific code)"
        "401":
          description: "Missing, invalid or expired credentials (empty body, WWW-Authenticate\
            \ header)"
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The caller's roles do not allow this operation (empty body),\
            \ or an operation-specific refusal (problem with a code)"
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Not found, or not visible to the caller's organization or\
            \ venue (code NOT_FOUND)"
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Rate limit exceeded (code RATE_LIMITED); see the Retry-After
            header
        "503":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "A dependency is unavailable (code DATABASE_UNAVAILABLE, STORAGE_UNAVAILABLE\
            \ or AUTHENTICATION_UNAVAILABLE); the request was not completed"
      security:
      - bearerAuth: []
      summary: List capture artifacts
      tags:
      - capture-controller
      x-chaya-roles:
      - ADMIN
      - VENUE_MANAGER
      - OPERATOR
    post:
      description: "Returns mediaId, partSizeBytes and totalParts."
      operationId: captureInitMedia
      parameters:
      - in: path
        name: venueId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: captureId
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InitResult"
          description: Created
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Malformed request: an unparseable parameter or body, or a\
            \ failed validation (code BAD_REQUEST or an operation-specific code)"
        "401":
          description: "Missing, invalid or expired credentials (empty body, WWW-Authenticate\
            \ header)"
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The caller's roles do not allow this operation (empty body),\
            \ or an operation-specific refusal (problem with a code)"
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Not found, or not visible to the caller's organization or\
            \ venue (code NOT_FOUND)"
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "CAPTURE_NOT_ACCEPTING_UPLOADS, TOO_MANY_FILES"
        "413":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "FILE_TOO_LARGE: larger than the limit for its kind, or DEVICE_METADATA_TOO_LARGE"
        "415":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "UNSUPPORTED_MEDIA_TYPE: content type not allowed"
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Rate limit exceeded (code RATE_LIMITED); see the Retry-After
            header
        "503":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "A dependency is unavailable (code DATABASE_UNAVAILABLE, STORAGE_UNAVAILABLE\
            \ or AUTHENTICATION_UNAVAILABLE); the request was not completed"
      security:
      - bearerAuth: []
      summary: Open a resumable multipart upload for one file
      tags:
      - capture-controller
      x-chaya-roles:
      - ADMIN
      - VENUE_MANAGER
      - OPERATOR
  /api/v1/venues/{venueId}/captures/{captureId}/media/{mediaId}:
    get:
      operationId: captureGetMedia
      parameters:
      - in: path
        name: venueId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: captureId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: mediaId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaView"
          description: OK
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Malformed request: an unparseable parameter or body, or a\
            \ failed validation (code BAD_REQUEST or an operation-specific code)"
        "401":
          description: "Missing, invalid or expired credentials (empty body, WWW-Authenticate\
            \ header)"
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The caller's roles do not allow this operation (empty body),\
            \ or an operation-specific refusal (problem with a code)"
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Not found, or not visible to the caller's organization or\
            \ venue (code NOT_FOUND)"
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Rate limit exceeded (code RATE_LIMITED); see the Retry-After
            header
        "503":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "A dependency is unavailable (code DATABASE_UNAVAILABLE, STORAGE_UNAVAILABLE\
            \ or AUTHENTICATION_UNAVAILABLE); the request was not completed"
      security:
      - bearerAuth: []
      summary: Media status and uploaded parts
      tags:
      - capture-controller
      x-chaya-roles:
      - ADMIN
      - VENUE_MANAGER
      - OPERATOR
  /api/v1/venues/{venueId}/captures/{captureId}/media/{mediaId}/complete:
    post:
      operationId: captureCompleteMedia
      parameters:
      - in: path
        name: venueId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: captureId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: mediaId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "202":
          description: Validation started
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Malformed request: an unparseable parameter or body, or a\
            \ failed validation (code BAD_REQUEST or an operation-specific code)"
        "401":
          description: "Missing, invalid or expired credentials (empty body, WWW-Authenticate\
            \ header)"
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The caller's roles do not allow this operation (empty body),\
            \ or an operation-specific refusal (problem with a code)"
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Not found, or not visible to the caller's organization or\
            \ venue (code NOT_FOUND)"
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "MISSING_PARTS, MEDIA_NOT_COMPLETABLE"
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Rate limit exceeded (code RATE_LIMITED); see the Retry-After
            header
        "503":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "A dependency is unavailable (code DATABASE_UNAVAILABLE, STORAGE_UNAVAILABLE\
            \ or AUTHENTICATION_UNAVAILABLE); the request was not completed"
      security:
      - bearerAuth: []
      summary: Assemble the parts and start asynchronous validation (also retries
        QUARANTINED)
      tags:
      - capture-controller
      x-chaya-roles:
      - ADMIN
      - VENUE_MANAGER
      - OPERATOR
...
components:
  schemas:
...
    CaptureView:
      type: object
      properties:
        acceptedMediaCount:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
        device:
          type: object
          additionalProperties: {}
        durationSeconds:
          type: number
          format: double
        endedAt:
          type: string
          format: date-time
        failureCode:
          type: string
        failureMessage:
          type: string
        floorId:
          type: string
          format: uuid
        id:
          type: string
          format: uuid
        mediaCount:
          type: integer
          format: int32
        operatorId:
          type: string
        qualityState:
          type: string
        startedAt:
          type: string
          format: date-time
        status:
          type: string
          enum:
          - CREATED
          - UPLOADING
          - UPLOADED
          - VALIDATING
          - READY_FOR_PROCESSING
          - PROCESSING
          - COMPLETED
          - FAILED
        venueId:
          type: string
          format: uuid
...
    CompleteUpload:
      type: object
      properties:
        durationSeconds:
          type: number
          format: double
        endedAt:
          type: string
          format: date-time
...
    CreateCapture:
      type: object
      properties:
        device:
          type: object
          additionalProperties: {}
        floorId:
          type: string
          format: uuid
        startedAt:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: "Keycloak access token for the chaya-api audience. Organization,\
        \ roles and venue memberships come from its claims."
      scheme: bearer
      type: http
    viewerToken:
      description: "Short-lived public-viewer token from POST /api/v1/public/viewer-token,\
        \ bound to one venue; acts as PUBLIC_VIEWER. Never together with a bearer\
        \ token (400)."
      in: header
      name: X-Chaya-Viewer-Token
      type: apiKey